+++ lib/graphql_rails/attributes/attribute.rb
@@ -13,14 +13,17 @@ module GraphqlRails
 
       attr_reader :attributes
 
-      def initialize(name, type = nil, description: nil, property: name, required: nil)
+      # rubocop:disable Metrics/ParameterLists
+      def initialize(name, type = nil, description: nil, property: name, required: nil, options: {})
         @initial_type = type
         @initial_name = name
+        @options = options
         @description = description
         @property = property.to_s
         @required = required
         @attributes ||= {}
       end
+      # rubocop:enable Metrics/ParameterLists
 
       def type(new_type = nil)
         return @initial_type if new_type.nil?
@@ -40,6 +43,13 @@ module GraphqlRails
         return @property if new_property.nil?
 
         @property = new_property.to_s
+        self
+      end
+
+      def options(new_options = {})
+        return @options if new_options.blank?
+
+        @options = new_options
         self
       end