+++ CHANGELOG.md
@@ -9,9 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added/Changed/Deprecated/Removed/Fixed/Security: YOUR CHANGE HERE
 
+## [1.2.0](2021-02-15)
+
+* Added/Changed/Deprecated/Removed/Fixed/Security: YOUR CHANGE HERE
+
 * Added: `options` argument to model level attribute. Allows disabling automatic camelCase
 * Fixed: methods with complex input arguments receives `Hash` instances instead of `GraphQL::Schema::InputObject`
 * Fixed: Using `ActiveSupport::ParameterFilter` (Rails 6.1), if it is defined, instead of `ActionDispatch::Http::ParameterFilter`
+* Changed: graphql version is now `1.12` which may require system-wide changes.
+* Fixed: improved connection wrapper for pagination to work.
+* Fixed: implementation of `total` field is no longer missing when using pagination.
+
 
 ## [1.0.0](2020-02-07)